split out NoteManager, SupplyRead, SupplyWrite traits
