+++ component/src/app/mod.rs
@@ -2,7 +2,7 @@ use std::sync::Arc;
 
 use anyhow::Result;
 use penumbra_chain::params::FmdParameters;
-use penumbra_chain::{genesis, StateReadExt as _, StateWriteExt as _};
+use penumbra_chain::{genesis, StateWriteExt as _};
 use penumbra_storage2::{AppHash, State, StateTransaction, StateWrite, Storage};
 use penumbra_transaction::Transaction;
 use tendermint::abci::{self, types::ValidatorUpdate};
@@ -12,7 +12,7 @@ use crate::dex::Dex;
 // use crate::governance::Governance;
 // use crate::ibc::IBCComponent;
 use crate::shielded_pool::ShieldedPool;
-// use crate::stake::component::Staking;
+use crate::stake::component::Staking;
 use crate::{Component, Context};
 
 pub mod state_key;
@@ -54,7 +54,7 @@ impl App {
         // The genesis block height is 0
         state_tx.put_block_height(0);
 
-        // Staking::init_chain(&mut state_tx, app_state).await;
+        Staking::init_chain(&mut state_tx, app_state).await;
         // IBCComponent::init_chain(&mut state_tx, app_state).await;
         Dex::init_chain(&mut state_tx, app_state).await;
         // Governance::init_chain(&mut state_tx, app_state).await;
@@ -75,7 +75,7 @@ impl App {
         // store the block time
         state_tx.put_block_timestamp(begin_block.header.time);
 
-        // Staking::begin_block(&mut state_tx, ctx.clone(), begin_block).await;
+        Staking::begin_block(&mut state_tx, ctx.clone(), begin_block).await;
         // IBCComponent::begin_block(&mut state_tx, ctx.clone(), begin_block).await;
         Dex::begin_block(&mut state_tx, ctx.clone(), begin_block).await;
         // Governance::begin_block(&mut state_tx, ctx.clone(), begin_block).await;
@@ -158,7 +158,7 @@ impl App {
     fn check_tx_stateless(ctx: Context, tx: Arc<Transaction>) -> Result<()> {
         // TODO: these can all be parallel tasks
 
-        // Staking::check_tx_stateless(ctx.clone(), tx)?;
+        Staking::check_tx_stateless(ctx.clone(), tx.clone())?;
         // IBCComponent::check_tx_stateless(ctx.clone(), tx)?;
         Dex::check_tx_stateless(ctx.clone(), tx.clone())?;
         // Governance::check_tx_stateless(ctx.clone(), tx)?;
@@ -175,7 +175,7 @@ impl App {
     ) -> Result<()> {
         // TODO: these can all be parallel tasks
 
-        // Staking::check_tx_stateful(state.clone(), ctx.clone(), tx.clone()).await?;
+        Staking::check_tx_stateful(state.clone(), ctx.clone(), tx.clone()).await?;
         // IBCComponent::check_tx_stateful(state.clone(), ctx.clone(), tx.clone()).await?;
         Dex::check_tx_stateful(state.clone(), ctx.clone(), tx.clone()).await?;
         // Governance::check_tx_stateful(state.clone(), ctx.clone(), tx.clone()).await?;
@@ -190,7 +190,7 @@ impl App {
         ctx: Context,
         tx: Arc<Transaction>,
     ) -> Result<()> {
-        // Staking::execute_tx(state, ctx.clone(), tx.clone()).await?;
+        Staking::execute_tx(state, ctx.clone(), tx.clone()).await?;
         // IBCComponent::execute_tx(state, ctx.clone(), tx.clone()).await?;
         Dex::execute_tx(state, ctx.clone(), tx.clone()).await?;
         // Governance::execute_tx(state, ctx.clone(), tx.clone()).await?;
