+++ component/src/shielded_pool/mod.rs
@@ -1,11 +1,13 @@
-mod commission;
 mod component;
 mod consensus_rules;
 pub(crate) mod event;
 mod metrics;
+mod note_manager;
+mod supply;
 
 pub mod state_key;
 
 pub use self::metrics::register_metrics;
-pub use commission::{CommissionAmount, CommissionAmounts};
 pub use component::{ShieldedPool, StateReadExt};
+pub use note_manager::NoteManager;
+pub use supply::{SupplyRead, SupplyWrite};
